requirements.txt
# Core Dependencies
flask==3.0.0
python-dotenv==1.0.0

# LLM & RAG
langchain==0.1.20
langchain-google-genai
langchain-community

# Vector Database
chromadb==0.4.24

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# Utilities
Werkzeug==3.0.1
google-generativeai